requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-appconf==1.0.2
django-auditlog==0.4.5
django-imagekit==4.0.2
django-jsonfield==1.0.1
djangorestframework==3.7.7
idna==2.6
pilkit==2.0
Pillow==5.0.0
pyfcm==1.4.4
PyMySQL==0.8.0
python-crontab==2.2.8
python-dateutil==2.6.0
pytz==2018.3
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
urllib3==1.22
social-auth-app-django==3.1.0
social-auth-core==3.2.0
reportlab==3.3.0